Refactored Service & added TaskService
- Service is no longer a child of Object
- Object::tick() now accepts a uint64_t of the UNIX EPOCH time in milliseconds
- Moved away from sf::Clock for our own callback based task scheduler
- Framerate is now unlocked and independent of physics
- TaskService has been added and can scheduler & maintain a queue of timers & scheduled tasks
- Root::init has been added and is required to startup our main tick task

+++ src/core/Entity.hpp
@@ -40,5 +40,5 @@ public:
     bool getAnchored();
     b2Body* getBody();
 
-    void tick(float);
+    void tick(uint64_t);
 };
